CORCI-448 scons check style and lint check
Check the style and lint for scons_local.
Run unit tests for scons_local.
Do test builds for CaRT, and components of IOF and DAOS.

Jenkinsfile is for testing daos-stack/pipeline-lib PR #8.

Jenkinsfile2 is for testing parameterized sconsbuild in that library.

Jenkinsfile3 is for testing running other distros and compilers as
additional child jobs.

Change-Id: Ice434d97635a3a267dad1573a98781878aaf02e9
Signed-off-by: John.E.Malmberg <john.e.malmberg@intel.com>